replace WhatIs to component
